requirements.txt
certifi==2019.11.28
chardet==3.0.4
coreschema==0.0.4
Django==2.2.7
django-rest-swagger==2.2.0
django-restframework==0.0.1
djangorestframework==3.10.3
idna==2.8
itypes==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
openapi-codec==1.3.2
Pillow==6.2.1
psycopg2-binary==2.8.4
pytz==2019.3
PyYAML==5.2
requests==2.22.0
simplejson==3.17.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.7
